register if not-registered OR skip-pass-void.
In probes.py, amend the decision logic for running registration.register().
Now assess existing reginfo.
If it's a skip, and the skip is for some OTHER version of LabGym, then ignore the skip and carry on with new registration.
(This consideration of skip_pass_void can be defeated, if desired, by changing the if-statement.)